// frontend/src/types/overtime-v2.ts
export enum StatusEnum {
  OPEN = 0,
  PAUSED = 1,
  RESOLVED = 10,
  CANCELLED = 255,
}
// ...
export type GameMarket = {
  gameId: string;
  subLeagueId: number;
  typeId: number;
  line: number;
  maturity: number;
  homeTeam: string;
  awayTeam: string;
  status: number;
  odds: number[];
  childMarkets: {
    typeId: number;
    line: number;
    status: number;
    odds: number[];
  }[];
  statusCode: string;
  tournamentName: string;
  marketTypeIds: number[];
  childMarketsCount: number;
};